Two different kNN algorithms
Now, Sent2VecManager.query(sentence, k, type="dist"), where 'type' can
be "dist" or "score".

In the case of "dist", scipy.spatial.distance.cdist is used to determine
distance between vectors (smaller distance is nearer).

In the case of "score", another approach using numpy.dot is used to
determine a score for each vector pair (higher is nearer).

The latter is the approach used in the original skip-thoughts implementation,
in particular, function "nn" in "skipthoughts.py".

The code is working.