Create Heroku app and specify Python dependencies in Pipfile
We use the official Heroku Python buildpack for our Python application:
/https://elements.heroku.com/buildpacks/heroku/heroku-buildpack-python

The Python buildpack installs pip and pipenv and supports the Pipfile
(https://github.com/pypa/pipfile) for retrieving Python dependencies.

That means, we specify all the dependencies for our Python application
in the Pipfile, and then the Python buildpack installs all these
dependencies with pip.

/// Skip-Thoughts Dependencies ///

Dependencies for skip-thoughts: https://github.com/ryankiros/skip-thoughts

However, we need only the "skipthoughts.py" script of this package, and
the dependencies for this file are:

- numpy
- scipy
- nltk
- Theano

Notes:

- scikit-learn is not required for "skipthoughts.py", but only for
other modules in the skipthoughts package that we don't need for the
Sent2Vec application.
- The installation page of Theano
(http://deeplearning.net/software/theano/install_ubuntu.html) talks
about libgpuarray, however, this is an optional dependency, and not
needed to use Theano.

Fortunately, all the above listed dependencies can be installed with pip.
So, we just list them in the Pipfile. Furthermore, skip-thoughts requires
Python 2.7, which we also specify in the Pipfile.

For now, we just want the latest version of each dependency, so we specify
'*' as the version for each package in the Pipfile.

The Python buildpack for Heroku supports a nltk.txt file for specifying
the NLTK Data corpora to download
(https://devcenter.heroku.com/articles/python-nltk). We need the 'punkt'
corpus for the sentence tokenisation.

/// Pipfile.lock ///

The Python buildpack recommends to manually generate the Pipfile.lock
(containing the exact versions of each package to use) and to check it
into Git and push it to Heroku.

To generate the Pipfile.lock, use "pipenv lock".

On Mac, apparently due to a bug, this might require to set the following
in ~/.bash_profile:

export LANG=en_US.UTF-8
export LC_ALL=en_US.UTF-8