Reorganisation and server script
Flatter organisation, modules instead of classes. The main.py file is
executable and starts the RPC server, decodes requests and encodes
responses, and carries out the Sent2Vec tasks through the
sent2vec_manager module.

The JSON message format that the server accepts follows the JSON-RPC
standard, however, it does not yet fully implement the specification.
The goal is to make this JSON-RPC 2.0 compliant in the future.

At the moment, request messages have the form:

{"method": X, "params": Y}

where X is a string (either "index" or "query") and Y is either a string
or an array of positional parameters (TODO: a string is not allowed in
the JSON-RPC 2.0 specification, this must be an array if the
"by-position" option is used).

And response messages have the form:

{"result": Z}

where Z is any type of object or null.