Create Docker image
Docker image has been created according to the Dockerfile and pushed to
Docker Hub (https://hub.docker.com/r/stormysmoke/sent2vec-back/).

The image includes the model files. Currently the pruned model, which is
900 MB, is included. For a production version of the image, the original
model, which is 5.3 GB, must be included in the model.

The idea is to make Docker image tags of the form "0.0.1-dev" and
"0.0.1", where the former images contain the pruned model (for
development only), and the latter images contain the original model (for
production).

The image uses the official Python image using Debian 9 (Stretch) as a
base. We can't easily use the much smaller Alpine version of the Python
image, because we need to install a dependency with apt-get for Theano,
and Alpine does not have apt-get.