Fix app to a working state
The app is now able again to handle requests from clients in the same
format as before. This is a quick fix, multi-user usage is not yet
included, and every index request from a client results in the indexed
text being overwritten on the server-side (as it was before the
rewrite).