Download models in Dockerfile rather than copying them from host
This has the following advantages:

- The large model files don't need to be sent to the Docker daemon at
the beginning of a build.
- Easier to switch to another model: just update the URLs in the
Dockerfile. Before, it was necessary to replace the existing model files
with the new model files in the appropriate directory.

In any case, the command that includes the model in the image takes a
long time (due to the large size of the model files): in one case it's
the download of the files from the Internet to the Docker image, in the
other case it's the copying of them from the host to the Docker image.

Because of this, one thing is very important: the command that includes
the image file in the image has to come very early in the Dockerfile so
that it is executed as rarely as possible. This is because Docker takes
layers (states after executing a Dockerfile command) from the build
cache, as long as no command has invalidated the cache. As soon as a
command invalidates the cache (because it was edited, or, if it's a RUN
command, a file it refers to has changed), then all the subsequent
Dockerfile commands are re-executed, even if nothing has changed with
them.

For example, if the command that includes the model comes after the
command that includes the Python files, then, if a Python file has been
edited, the model-including command is re-executed when re-building the
image, even if nothing has changed with the model command itself. This
can be avoided by simply putting the model-including command before the
Python includig command in the Dockerfile.